04_Base_Datos/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
